scripts/python/codex/requirements.txt
# Dependências do Codex Chat (GPT-5)
openai>=1.51.0
python-dotenv>=1.0.1
